survey-data-dictionary: Fix our use of Node + TypeScript runtime
It was broken by a change to the vendor source code which require()-ed a
PNG file.

The key fix is the addition of a require() handler for .png files.  Due
to quirks of ts-node, however, that necessitated a switch from the
ts-node CLI to loading the ts-node module from within Node.

This fixes the Node runtime issue, but the survey parsing is still
broken because of data changes.  Those are to be addressed shortly.