survey-flow: More robustly test branching conditions
While nothing was broken yet, it is more robust to test full set
equality of the branch conditions instead of using a cascade of element
membership tests.  For example, if future branches allow more
multi-condition combinations (in addition to the existing choice +
location combo), this code will now hit the guard assertion and fail
instead of silently doing the wrong thing.  This also means that the
order of the if/elif branches is no longer as significant.