Change immunization to questionnaire response
The Immunization resource references the Encounter where the
immunization took place, not where the Immunization was reported, so
we are changing it to a QuestionnaireResponse.item. There can be
two answers to the same question, one representing whether an
immunization was done and what date it was done if known.