SCAN: Document return of results portal
We often get requests for screenshots of the latest views on the
SCAN SecureLink results portal and the related results PDFs. Instead of
manually curating these screenshots and sending over PDFs for each new
request, point study members to the live development instance of the
SecureLink site where the test data are accessible.