SCAN: Update RoR documentation to reflect rebranding
We are rebranding the way we return results to participants, moving SCAN
under the SFS umbrella. Update some of our language plus direct links to
the scripts used to return results to participants.

Hold off for now on renaming the entire SCAN documentation page until we
have a better idea of the crossover and similarities in SCAN vs other
SFS studies once UW Reopening launches.