Linelists: Ephasize updating environment each run
This step is easily overlooked with the way the documentation is
currently written, and not syncing the development environment with the
latest changes causes frequent bugs (and headaches) for our users.

Duplicate the warning message about updating the python environment. I
prefer this approach to duplicating the code itself, as there is a
greater chance in the latter scenario that the duplicated instructions
may diverge.